requirements.txt
aiogram==3.15.0
python-dotenv==1.0.1
aiosqlite==0.20.0
aiohttp_socks==0.9.2
fpdf2==2.8.2
